backend_examples/requirements.txt
# Django and Core
Django>=4.2.0,<5.0
python-decouple>=3.8

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Django Channels (WebSocket support)
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Redis (for Channel Layer)
redis>=5.0.0
hiredis>=2.2.0

# CORS
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL (recommended)
# mysqlclient>=2.2.0    # MySQL (alternative)

# File handling
Pillow>=10.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development
django-debug-toolbar>=4.2.0
ipython>=8.12.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Monitoring (optional)
sentry-sdk>=1.32.0
